rename path-changes.js to change-recorder.js
